fix(eq-f90): GRal/GZal/GAlpRP are default REAL, not REAL(8)
Bugbot MED on PR #71: G-starting variables are not covered by the
shim IMPLICIT REAL*8(A,B,D-F,H,O-Z) rule, so the original COMMON
code treated them as default REAL (typically REAL*4). Declaring
them REAL(8) would change storage size and break binary layout for
consumers that access them via the eqcomq/eqcomm shim.

+++ eq/eqcom3_mod.f90
@@ -81,8 +81,13 @@ MODULE eqcom3_mod
   ! --- ripple grids ---
   REAL(8) :: Rrp(NRrpM), Zrp(NZrpM), RpplRZ(NRrpM, NZrpM)
   REAL(8) :: URpplRZ(4, 4, NRrpM, NZrpM)
-  REAL(8) :: GRal(NRM, NTVM), GZal(NRM, NTVM)
-  REAL(8) :: GAlpRP(NRM, NTVM)
+  ! NOTE: G-starting variables fall OUTSIDE the shim IMPLICIT REAL*8
+  ! rule (A,B,D-F,H,O-Z), so the original COMMON-based code treated
+  ! them as default REAL (REAL*4). Declaring REAL(8) here would change
+  ! storage size and break binary layout for downstream consumers
+  ! reading via the eqcomq/eqcomm shim.
+  REAL :: GRal(NRM, NTVM), GZal(NRM, NTVM)
+  REAL :: GAlpRP(NRM, NTVM)
   INTEGER :: NtrcMAX(NRM)
 
 END MODULE eqcom3_mod
